feat(frontend): phase 12 — file-rename BatchContext.tsx → ProgramContext.tsx
frontend/src/context/ProgramContext.tsx
  - The canonical implementation now lives here. The
    full ProgramContext (types, hooks, provider, legacy
    aliases) moved to this file. New code imports from
    '../context/ProgramContext'.

frontend/src/context/BatchContext.tsx
  - Now a 16-line re-export shim. Every existing
    'from ../context/BatchContext' import keeps working
    unchanged — the file re-publishes the same symbols
    from ProgramContext.
  - When v1.70 ships the cutover, this shim can be
    deleted (and the ~30 consumer imports updated to the
    new path) in a single follow-up commit.

The full rename is now done in two steps:
  1. v1.69 (previous turn) — additive rename, every old
     export kept as a delegate to the new export.
  2. v1.69 (this turn) — file-level rename with a shim.
     No call sites had to change; tsc still clean across
     the whole frontend.

Refs: context/multi-program-cms-design.md — phase 12
frontend refactor (file rename).
